+++ calculater2.js
@@ -17,8 +17,8 @@ rl.on("line", (line) => {
 });
  
 rl.on('close', () => {
-    let first = parseInt(input[0]);
-    let second = parseInt(input[2]);
+    let first = parseFloat(input[0]);
+    let second = parseFloat(input[2]);
     if(input[1] === '+'){
         let plus = first + second;
         console.log(`result : ${plus}`);
